Extract to new Connection module
